+++ packages/bpk-component-accordion/src/BpkAccordionItem.js
@@ -16,12 +16,14 @@
  * limitations under the License.
  */
 
+/* @flow strict */
+
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, { type Node, type Element } from 'react';
 import AnimateHeight from 'bpk-animate-height';
 import { withButtonAlignment } from 'bpk-component-icon';
 import ChevronDownIcon from 'bpk-component-icon/sm/chevron-down';
-import BpkText from 'bpk-component-text';
+import BpkText, { TEXT_STYLES } from 'bpk-component-text';
 import { cssModules } from 'bpk-react-utils';
 
 import STYLES from './BpkAccordionItem.scss';
@@ -30,7 +32,18 @@ const getClassName = cssModules(STYLES);
 
 const ExpandIcon = withButtonAlignment(ChevronDownIcon);
 
-const BpkAccordionItem = props => {
+type Props = {
+  children: Node,
+  id: string,
+  title: string,
+  expanded: boolean,
+  icon: ?Element<any>,
+  onClick: () => mixed,
+  tagName: 'span' | 'p' | 'text' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6',
+  textStyle: $Keys<typeof TEXT_STYLES>,
+};
+
+const BpkAccordionItem = (props: Props) => {
   const iconClassNames = [getClassName('bpk-accordion__item-expand-icon')];
   const {
     id,
@@ -47,6 +60,7 @@ const BpkAccordionItem = props => {
   // if this component is passed initiallyExpanded, this makes sure it doesn't
   // end up on the node. Not ideal as our container component shouldn't be passing
   // it, but the benefit of a better container api versus this was worth it
+  // $FlowFixMe - see above
   delete rest.initiallyExpanded;
 
   if (expanded) {
@@ -63,6 +77,7 @@ const BpkAccordionItem = props => {
     : null;
 
   return (
+    // $FlowFixMe - inexact rest. See decisions/flowfixme.md
     <div id={id} {...rest}>
       <dt
         aria-level="3"
@@ -120,7 +135,7 @@ BpkAccordionItem.defaultProps = {
   icon: null,
   onClick: () => null,
   tagName: 'span',
-  textStyle: 'base',
+  textStyle: TEXT_STYLES.base,
 };
 
 export default BpkAccordionItem;